% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmte_aalen.R
\name{gmte_aalen}
\alias{gmte_aalen}
\title{gmte_aalen}
\usage{
gmte_aalen(
  Y_t0,
  Y_t1,
  Y_d,
  T,
  G,
  Z,
  D,
  Nsim = 100,
  alpha = 0.05,
  doCAT = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y_t0}{Variable name (string) for when participants "enter" the model, which appears in data.frame \code{D}. When participants enter the model (can be all 0s if \code{Y_t1} is time since start of exposure). Variable can be in date format from the \code{as.Date()} function, or numeric.}

\item{Y_t1}{Variable name (string) for when participants "exit" the model, which appears in data.frame \code{D}. Either days since start of exposure period (numeric) or in date format from the \code{as.Date()} function.}

\item{Y_d}{Variable name for the binary "event" variable (string) which appears in data.frame \code{D}.}

\item{T}{The treatment variable name (string) which appears in data.frame \code{D}. Assumed to be binary.}

\item{G}{The genotype variable name (string) which appears in data.frame \code{D}. Normally binary (e.g. comparing homozygous rare individuals to the rest of the population) but can be additive (0, 1, 2) or a score of multiple variants if desired.}

\item{Z}{A string containing the model covariates to appear in the \code{glm()} models (for example "age+sex"). All need to be in data.frame \code{D}.}

\item{D}{A data.frame containing the above variables.}

\item{Nsim}{Number of simulations to perform in the \code{aalen()} models. Default is 100.}

\item{alpha}{The p-value threshold for the chi-square test, estimating whether two estimates should be combined. Default is 0.05.}

\item{doCAT}{Run the CAT analysis? Default is FALSE.}

\item{verbose}{Return lots of output - useful for error checking. Default is FALSE.}
}
\value{
An object of class \code{twistR_GMTE} containing the following components:\describe{
\item{\code{CAT}}{The summary statistics from the Corrected As Treated (CAT) analysis.}
\item{\code{GMTE1}}{The summary statistics from the GMTE(1) analysis (i.e. in the treated individuals).}
\item{\code{GMTE0}}{The summary statistics from the GMTE(0) analysis (i.e. in the untreated individuals).}
\item{\code{MR}}{The summary statistics from the MR analysis.}
\item{\code{RGMTE}}{The summary statistics from the Robust GMTE analysis (GMTE1 corrected for GMTE0).}
\item{\code{FullCombined}}{The combined summary statistics from all analyses performed, inclduing combinations.}
}
}
\description{
Performs analyses for the Triangulation WIthin A STudy (TWIST) framework to calcuate the ‘genetically moderated treatment effect’ (GMTE) for a time-to-event Aalen additive hazards model. The \code{gmte_aalen} function returns an object of class \code{twistR_GMTE}, containing effect estimates from the individual tests (such as RGMTE) and the results when combinations are performed (such as RGMTE+MR).
}
\examples{
# Example using a time-to-event outcome (mortality), binary treatment (statins), and binary genotype (SLCO1B1*5 homozygotes) variables
Y_t0="date_first_statin"
Y_t1="date_of_death_or_censor"
Y_d="dead"
T="statin"
G="slco1b1_5_hmz"
Z="age+PC1+PC2+PC3+PC4+PC5+PC6+PC7+PC8+PC9+PC10"
results=gmte_aalen(Y_t0,Y_t1,Y_d,T,G,Z,D)
}
\references{
Bowden, J., et al., The Triangulation WIthin A STudy (TWIST) framework for causal inference within Pharmacogenetic research. PLoS Genetics. https://doi.org/10.1371/journal.pgen.1009783
}
\author{
Jack Bowden; Luke Pilling.
}
