% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmte_binary.R
\name{gmte_binary}
\alias{gmte_binary}
\title{gmte_binary}
\usage{
gmte_binary(
  Y,
  T,
  G,
  Z,
  D,
  Link = "logit",
  alpha = 0.05,
  doCAT = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{The binary outcome variable name (string) which appears in data.frame \code{D}.}

\item{T}{The treatment variable name (string) which appears in data.frame \code{D}. Assumed to be binary.}

\item{G}{The genotype variable name (string) which appears in data.frame \code{D}. Normally binary (e.g. comparing homozygous rare individuals to the rest of the population) but can be additive (0, 1, 2) or a score of multiple variants if desired.}

\item{Z}{A string containing the model covariates to appear in the \code{glm()} models (for example "age+sex"). All need to be in data.frame \code{D}.}

\item{D}{A data.frame containing the above variables.}

\item{Link}{Link function for the \code{glm()} - needs to be one of "logit","probit" or "identity". If unspecified the default is "logit".}

\item{alpha}{The p-value threshold for the chi-square test, estimating whether two estimates should be combined. Default is 0.05.}

\item{doCAT}{Run the CAT analysis? Default is FALSE.}

\item{verbose}{Return lots of output - useful for error checking. Default is FALSE.}
}
\value{
An object of class \code{twistR_GMTE} containing the following components:\describe{
\item{\code{CAT}}{The summary statistics from the Corrected As Treated (CAT) analysis.}
\item{\code{GMTE1}}{The summary statistics from the GMTE(1) analysis (i.e. in the treated individuals).}
\item{\code{GMTE0}}{The summary statistics from the GMTE(0) analysis (i.e. in the untreated individuals).}
\item{\code{MR}}{The summary statistics from the MR analysis.}
\item{\code{RGMTE}}{The summary statistics from the Robust GMTE analysis (GMTE1 corrected for GMTE0).}
\item{\code{FullCombined}}{The combined summary statistics from all analyses performed, inclduing combinations.}
}
}
\description{
Performs analyses for the Triangulation WIthin A STudy (TWIST) framework to calcuate the ‘genetically moderated treatment effect’ (GMTE) for a binary outcome. The \code{gmte_binary} function returns an object of class \code{twistR_GMTE}, containing effect estimates from the individual tests (such as RGMTE) and the results when combinations are performed (such as RGMTE+MR).
}
\examples{
# Example using a binary outcome (high LDL), binary treatment (statins), and binary genotype (SLCO1B1*5 homozygotes) variables
Y="ldl_high"
T="statin"
G="slco1b1_5_hmz"
Z="age+PC1+PC2+PC3+PC4+PC5+PC6+PC7+PC8+PC9+PC10"
Link="logit"
results=gmte_binary(Y,T,G,Z,D,Link)
}
\references{
Bowden, J., et al., The Triangulation WIthin A STudy (TWIST) framework for causal inference within Pharmacogenetic research. PLoS Genetics. https://doi.org/10.1371/journal.pgen.1009783
}
\author{
Jack Bowden; Luke Pilling.
}
