% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmte_combine.R
\name{gmte_combine}
\alias{gmte_combine}
\title{gmte_combine}
\usage{
gmte_combine(Ests = Ests, SEs = SEs, alpha = 0.05)
}
\arguments{
\item{Ests}{The estimates.}

\item{SEs}{Standard Errors for the estimates.}

\item{alpha}{Significance threshold (default 0.05)}
}
\value{
An object containing the combined estimate (plus SE, P, and Q statistic)
}
\description{
Internal function to combine estimates from multiple models and provide summary statistics
}
\references{
Bowden, J., et al., The Triangulation WIthin A STudy (TWIST) framework for causal inference within Pharmacogenetic research. PLoS Genetics. https://doi.org/10.1371/journal.pgen.1009783
}
\author{
Jack Bowden; Luke Pilling.
}
